requirements.txt
flask
flask-cors
Flask-SQLAlchemy
Flask-Migrate
SQLAlchemy-serializer
python-dotenv
rq
redis
uWSGI